feat(P1-M4-A): asset database hardening - 64-bit IDs, metadata, thumbnails
P1-M4-A1: 64-bit Asset ID Hashing
- Replace 32-bit FNV with 64-bit FNV-1a for asset IDs
- Update AssetId typedef from uint32_t to uint64_t across all modules
- Add uint64 support to JSON writer/parser (write_uint64, as_uint64)
- Fix Lua mesh ID range check for 64-bit values
- Migrate asset packer metadata to schema version 2 with 64-bit IDs
- Add 100K hash collision test (zero collisions verified)

P1-M4-A2: Asset Metadata System
- Create AssetMetadata struct with type tags, import settings, tags[]
- Enhanced .meta.json sidecar with assetId, typeTag, fileSize, importSettings
- Tag system with add/has/query API (queryable from editor and runtime)
- Per-asset-type import settings (mesh: scale/upAxis/normals, texture: format/mips/sRGB)

P1-M4-A3: Asset Thumbnails
- CPU software rasterizer for 64x64 mesh thumbnails (no GL dependency)
- Thumbnail mtime invalidation (skip regeneration if thumbnail newer than source)
- Editor thumbnail loading and display via ImGui::Image in asset browser
- Thumbnails stored as PNG in .thumbnails/ directory

All 56 tests pass. Zero warnings.

+++ runtime/include/engine/runtime/world.h
@@ -102,7 +102,7 @@ struct ScriptComponent final {
 
 // Renderer-facing component; keep minimal to avoid bloating draw commands.
 struct MeshComponent final {
-  std::uint32_t meshAssetId = 0U;
+  std::uint64_t meshAssetId = 0ULL;
   math::Vec3 albedo = math::Vec3(1.0F, 1.0F, 1.0F);
   float roughness = 0.5F;
   float metallic = 0.0F;
